Updated users endpoint

+++ routes/users.js
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 import User from '../models/user.js';
-import { error, getLastId } from '../utils.js';
+import { error, getLastId, validToken } from '../utils.js'
+import { genSalt, hash } from 'bcryptjs'
 const router = Router();
 
 router.get('/', async (req, res) => {
@@ -26,19 +27,78 @@ router.post('/', async (req, res) => {
 });
 
 router.put('/:id', async (req, res) => {
+  const user = await validToken(req, res)
+  if (user === null) return
+
+  if (user.role.name !== 'admin' && user.id !== req.params.id) {
+    error(res, 'no-permission', 404);
+    return
+  }
+
+  if (await User.findOne({
+    _id: {
+      $ne: user._id,
+    },
+    email: req.body.email
+  })) {
+    error(res, 'email.already-used', 404);
+    return
+  }
+
   const item = await User.findOneAndUpdate(
     { _id: req.params.id },
-    req.body,
-    { new: true }
-  );
-  if (!item) {
-    error(res, 'user.not-found', 404);
-    return;
-  }
+    {
+      $set: {
+        firstname: req.body.firstname,
+        name: req.body.name,
+        email: req.body.email,
+        description: req.body.description,
+        birthday: req.body.birthday,
+      }
+    },
+  )
   res.json(item);
 });
 
+router.put('/:id/password', async (req, res) => {
+  const user = await validToken(req, res)
+  if (user === null) return
+
+  if (user.role.name !== 'admin' && user.id !== req.params.id) {
+    error(res, 'no-permission', 404);
+    return
+  }
+
+  let email = user.email;
+  if (user.id !== req.params.id) {
+    email = (await User.findOne({_id: req.params.id}, "email")).email
+  }
+
+  const salt = await genSalt()
+  const password = await hash(req.body.password, salt)
+  const token = await hash(email + req.body.password + Date.now(), salt);
+
+  await User.updateOne(
+    { _id: req.params.id },
+    {
+      $set: {
+        password: password,
+        token: token
+      }
+    },
+  )
+  res.json(token);
+});
+
 router.delete('/:id', async (req, res) => {
+  const user = await validToken(req, res)
+  if (user === null) return
+
+  if (user.role.name !== 'admin' && user.id !== req.params.id) {
+    error(res, 'no-permission', 404);
+    return
+  }
+
   const item = await User.findOneAndDelete({ _id: req.params.id });
   if (!item) {
     error(res, 'user.not-found', 404);
